+++ src/CSDP.jl
@@ -5,9 +5,6 @@
 
 module CSDP
 
-import LinearAlgebra
-import SparseArrays
-
 import CSDP_jll
 
 function __init__()
@@ -25,21 +22,12 @@ const CSDP_INT = Cint
 
 CSDP uses 1-based indexing for its arrays. But since C has 0-based indexing, all
 1-based vectors passed to CSDP need to be padded with a "0'th" element that will
-never be accessed. To avoid doing this padding in Julia, we convert the vector
-to a reference, and use the optional second argument to ensure the reference
-points to the "0'th" element of the array. This is safe to do, provided C never
-accesses `x[0]`.
+never be accessed.
 """
 offset(x::Vector{T}) where {T} = pointer(x) - sizeof(T)
-# offset(x::Vector) = Ref(x, 0)
 
-include("blockmat.h.jl")
-include("blockmat.jl")
-include("declarations.h.jl")
+include("c_api.jl")
 include("MOI_wrapper.jl")
-
-export initsoln, easy_sdp
-export read_prob
-export BlockMatrix, ConstraintMatrix
+include("simple_api.jl")
 
 end # module
